Rename plugin class for Composer compatibility
The RoundCube plugin installer uses the Composer package name of the
plugin and does not support alternative names. This breaks the plugin,
as folder name and plugin class name differ and RoundCube can no longer
load the plugin. To allow Composer-based installations, the class name
was updated to match the plugin package name.